tpo-bd2/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==4.3.0
cassandra-driver==3.25.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
fastapi==0.115.13
geomet==0.2.1.post1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
redis==6.2.0
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
